major, i'm burning up - medigun categorization jumpscare (#21855)
* deepwoken

* applying suggestions, thank you softcerv


* drops a filler line

* where do these descs even show up

---------

+++ modular_skyrat/modules/customization/game/objects/items/conversion_kits.dm
@@ -8,16 +8,22 @@
 	/// The object to turn it into.
 	var/obj/item/to_obj
 
-/obj/item/device/custom_kit/afterattack(obj/target_obj, mob/user as mob, proximity_flag)
+/obj/item/device/custom_kit/afterattack(obj/target_obj, mob/user, proximity_flag)
 	if(!proximity_flag) //Gotta be adjacent to your target
 		return
 	if(isturf(target_obj)) //This shouldn't be needed, but apparently it throws runtimes otherwise.
 		return
 	else if(target_obj.type == from_obj) //Checks whether the item is eligible to be converted
+		if(!pre_convert_check(target_obj, user))
+			return FALSE
 		var/obj/item/converted_item = new to_obj(get_turf(src))
 		user.put_in_hands(converted_item)
 		user.visible_message(span_notice("[user] modifies [target_obj] into [converted_item]."), span_notice("You modify [target_obj] into [converted_item]."))
 		qdel(target_obj)
 		qdel(src)
 	else
-		user.visible_message(span_warning("It looks like this kit won't work on [target_obj]..."))
+		to_chat(user, span_warning("It looks like this kit won't work on [target_obj]..."))
+
+/// Override this if you have some condition you want fulfilled before allowing the conversion. Return TRUE to allow it to convert, return FALSE to prevent it.
+/obj/item/device/custom_kit/proc/pre_convert_check(obj/target_obj, mob/user)
+	return TRUE
